reformat to have multiple scripts
